ssh: explain why we need to pass to xprog absolute paths
"go test" will execute xprog in the directory (or directories) corresponding to the package(s) specified to the "go test" invocation. For example:

    $ go test -exec="xprog ssh --cfg $PWD/ssh_config.vagrant --" ./foo

will run xprog in directory `./foo`. This is why it is important to specify the ssh_config with an absolute path: `--cfg $PWD/ssh_config.vagrant`, so that it will be found no matter the xprog working directory.